backend/requirements.txt
asgiref
certifi==2020.12.5
chardet==4.0.0
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-sendsms==0.4
django-simple-history==2.12.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-contact==0.0.2
drf-transaction==0.0.3
drfaddons==0.1.4
idna==2.10
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
django-simple-history
pymongo==3.7.2
djongo==1.3.4
gunicorn==20.0.4
